# Sparse get dummies perf