@pep8speaks 2018-07-20
Hello @TomAugspurger! Thanks for submitting the PR.

 - In the file [`asv_bench/benchmarks/reshape.py`](https://github.com/pandas-dev/pandas/blob/8fda4fb9f63f30d4640cb5a134fce1ce88f06b53/asv_bench/benchmarks/reshape.py), following are the PEP8 issues :

> [Line 140:1](https://github.com/pandas-dev/pandas/blob/8fda4fb9f63f30d4640cb5a134fce1ce88f06b53/asv_bench/benchmarks/reshape.py#L140): [E303](https://duckduckgo.com/?q=pep8%20E303) too many blank lines (3)